requirements.txt

google-adk>=1.0.0
python-dotenv>=0.19.0
flask-cors>=3.0.10
uuid>=1.30
google-generativeai>=0.3.0
asgiref>=3.7.0

